fix(pinned-searches): Delete ALL groupsearchviews upon savedsearch deletion (#72779)
We need to ensure that users' saved searches are persisted as custom
views, AND that those users with a persisted saved saerch also have a
second "Priorizted" tab. This PR updates the pinned searches endpoint to
maintain this consistency by deleting ALL groupsearchviews upon
savedsearch deletion (as opposed to only the groupsearchview associated
with the user's saved search), and adding/updating the "Prioritized"
view when a saved search is added/updated.


This needs to be merged before the savedsearch -> groupsearchview
[backfill](https://github.com/getsentry/sentry/pull/71731) is run

+++ src/sentry/api/endpoints/organization_pinned_searches.py
@@ -65,6 +65,20 @@ class OrganizationPinnedSearchEndpoint(OrganizationEndpoint):
                 "query_sort": result["sort"],
             },
         )
+        # These groupsearchview entries are temporarily here to ensure that pinned searches
+        # are being dynamically upgraded to custom views until custom views are GA'd, at which
+        # point saved searches will be removed entirely, along with this endpoint.
+        GroupSearchView.objects.create_or_update(
+            organization=organization,
+            user_id=request.user.id,
+            position=1,
+            values={
+                "name": "Prioritized",
+                "query": "is:unresolved issue.priority:[high, medium]",
+                "query_sort": SortOptions.DATE,
+            },
+        )
+
         pinned_search = SavedSearch.objects.get(
             organization=organization,
             owner_id=request.user.id,
@@ -85,7 +99,5 @@ class OrganizationPinnedSearchEndpoint(OrganizationEndpoint):
             type=search_type.value,
             visibility=Visibility.OWNER_PINNED,
         ).delete()
-        GroupSearchView.objects.filter(
-            organization=organization, user_id=request.user.id, position=0
-        ).delete()
+        GroupSearchView.objects.filter(organization=organization, user_id=request.user.id).delete()
         return Response(status=204)
